feat: Refactor events page & components to add live sensors data
- Refactor Sensors.js into separate child components, util files
- Refactor helper functions in Sensors.js to use React hooks
- Remove unused table columns, mock data generation
- Add styling for live sensor data, including setting table to have min-height
