+++ web/client/src/App.scss
@@ -21,6 +21,7 @@ $events-map-mobile-threshold: 1100px;
 @import './components/EarthquakeList/EarthquakeList.scss';
 @import './components/SaveHeader/SaveHeader.scss';
 @import './components/Field/Field.scss';
+@import './components/SensorsMap/SensorsMap.scss';
 
 @import './scrollbar.scss';
