+++ web/client/src/components/SensorsMap/SensorsMap.scss
@@ -1,0 +1,3 @@
+.map-wrapper.mapboxgl-map {
+  background-color: #343332;
+}
