+++ web/client/src/components/SensorsTable/headers.js
@@ -1,0 +1,36 @@
+const headers = [
+  {
+    header: 'Sensor ID',
+    key: 'id',
+  },
+  {
+    header: 'Last report',
+    key: 'lastCheckin',
+  },
+  // {
+  //   header: 'Country',
+  //   key: 'country',
+  // },
+  {
+    header: 'Connection type',
+    key: 'connection_type',
+  },
+  {
+    header: 'Device version',
+    key: 'firmwareVer',
+  },
+  // {
+  //   header: 'Peak acc.',
+  //   key: 'peak_acc',
+  // },
+  {
+    header: 'Coordinates',
+    key: 'pos',
+  },
+  // {
+  //   header: 'Earthquakes detected',
+  //   key: 'earthquakes_detected',
+  // },
+]
+
+export default headers
