+++ web/client/src/content/Dashboard/Dashboard.scss
@@ -1,5 +1,4 @@
 .dashboard__container {
-  transition: margin-left 0.12s cubic-bezier(0.2, 0, 1, 0.9);
   margin-top: $layout-05;
   margin-left: $layout-01;
