+++ web/client/src/content/Sensors/dropdownItems.js
@@ -1,0 +1,10 @@
+export const sensorGroupDropdownItems = [
+  {
+    id: 'network-sensors',
+    text: 'Network sensors',
+  },
+  {
+    id: 'my-sensors',
+    text: 'My sensors',
+  },
+]
